model-service/requirements.txt
boto3==1.40.4
fastapi==0.116.1
huggingface_hub==0.34.3
locust==2.34.0
Pillow==11.3.0
pydantic==2.11.7
pydantic_settings==2.10.1
python-multipart==0.0.20
torch==2.8.0
transformers==4.55.2
typer==0.16.0
uvicorn[standard]==0.35.0
